+++ src/lib/Scenes/Artwork/Components/__tests__/ArtworkExtraLinks-tests.tsx
@@ -14,7 +14,13 @@ describe("ArtworkExtraLinks", () => {
   it("redirects to consignments flow when consignments link is clicked", () => {
     const component = mount(
       <Theme>
-        <ArtworkExtraLinks consignableArtistsCount={3} artistName={null} />
+        <ArtworkExtraLinks
+          artworkSlug={"123"}
+          isAcquireable={false}
+          isInquireable={false}
+          consignableArtistsCount={3}
+          artistName={null}
+        />
       </Theme>
     )
     const consignmentsLink = component.find(Text).at(1)
@@ -27,7 +33,13 @@ describe("ArtworkExtraLinks", () => {
     it("shows plural link text", () => {
       const component = mount(
         <Theme>
-          <ArtworkExtraLinks consignableArtistsCount={3} artistName={null} />
+          <ArtworkExtraLinks
+            artworkSlug={"123"}
+            isAcquireable={false}
+            isInquireable={false}
+            consignableArtistsCount={3}
+            artistName={null}
+          />
         </Theme>
       )
       expect(component.text()).toContain("Want to sell a work by these artists?")
@@ -35,7 +47,13 @@ describe("ArtworkExtraLinks", () => {
     it("shows consign link if at least 1 artist is consignable", () => {
       const component = mount(
         <Theme>
-          <ArtworkExtraLinks consignableArtistsCount={3} artistName={null} />
+          <ArtworkExtraLinks
+            artworkSlug={"123"}
+            isAcquireable={false}
+            isInquireable={false}
+            consignableArtistsCount={3}
+            artistName={null}
+          />
         </Theme>
       )
       expect(component.text()).toContain("Consign with Artsy.")
@@ -43,7 +61,13 @@ describe("ArtworkExtraLinks", () => {
     it("doesn't render component if no artists are consignable", () => {
       const component = mount(
         <Theme>
-          <ArtworkExtraLinks consignableArtistsCount={0} artistName={null} />
+          <ArtworkExtraLinks
+            artworkSlug={"123"}
+            consignableArtistsCount={1}
+            isAcquireable={false}
+            isInquireable={false}
+            artistName={null}
+          />
         </Theme>
       )
       expect(component).toEqual({})
@@ -54,7 +78,13 @@ describe("ArtworkExtraLinks", () => {
     it("shows singular link text", () => {
       const component = mount(
         <Theme>
-          <ArtworkExtraLinks consignableArtistsCount={1} artistName="Santa Claus" />
+          <ArtworkExtraLinks
+            artworkSlug={"123"}
+            isAcquireable={false}
+            isInquireable={false}
+            consignableArtistsCount={1}
+            artistName="Santa Claus"
+          />
         </Theme>
       )
       expect(component.text()).toContain("Want to sell a work by Santa Claus?")
@@ -63,10 +93,49 @@ describe("ArtworkExtraLinks", () => {
     it("shows consign link", () => {
       const component = mount(
         <Theme>
-          <ArtworkExtraLinks consignableArtistsCount={1} artistName="Santa Claus" />
+          <ArtworkExtraLinks
+            artworkSlug={"123"}
+            isAcquireable={false}
+            isInquireable={false}
+            consignableArtistsCount={1}
+            artistName="Santa Claus"
+          />
         </Theme>
       )
       expect(component.text()).toContain("Consign with Artsy.")
+    })
+  })
+  describe("FAQ and specialist links", () => {
+    it("renders FAQ link when isInquireable", () => {
+      const component = mount(
+        <Theme>
+          <ArtworkExtraLinks
+            artworkSlug={"123"}
+            isAcquireable={false}
+            isInquireable={true}
+            consignableArtistsCount={1}
+            artistName="Santa Claus"
+          />
+        </Theme>
+      )
+      expect(component.text()).toContain("Read our FAQ")
+      expect(component.text()).not.toContain("ask a specialist")
+    })
+
+    it("renders ask a specialist link when isAcquireable", () => {
+      const component = mount(
+        <Theme>
+          <ArtworkExtraLinks
+            artworkSlug={"123"}
+            isAcquireable={true}
+            isInquireable={true}
+            consignableArtistsCount={1}
+            artistName="Santa Claus"
+          />
+        </Theme>
+      )
+      expect(component.text()).toContain("Read our FAQ")
+      expect(component.text()).toContain("ask a specialist")
     })
   })
 })