+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 
 - Fixes intermittent crash on opening message attachment previews - ash
 - Fix consignments artist search and make it use relay - alloy
+- Adds links to Artwork view for FAQ and Ask a Specialist - kierangillen
+
+### Master
+
 - Adds Commercial Bid Buttons - kierangillen
 
 ### 1.12.16